Experiment: WebGL support
This enables (optional) WebGL support on the tile-based layers. If
WebGL is disabled, it falls back to normal rendering. Vector layers, as
well as ImageLayer-based rendering (used for untiled WMS), still use
canvas.

In my experience, this changes the usability of the viewer on my phone
running Firefox from "bearably slow", to "buttery smooth 90Hz". However,
I don't think this is usable as-is, due to an OpenLayers issue.

Compared to the canvas renderer, if a WebGL tile is partially
overlapped due to only some of the tiles of a higher zoom level being
loaded, it's not properly clipped but instead overlaps the two tiles,
showing both zoom levels simultaneously for transparent layers. Fixing
this would require a small-ish patch to OpenLayers.